pyproject.toml
[tool.poetry]
name = "tickler"
version = "0.1.0"
description = ""
authors = ["Leo Kukharau <kuhorev.leonid@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.3"
djangorestframework = "^3.12.2"
django-cors-headers = "^3.5.0"
django-environ = "^0.4.5"
psycopg2-binary = "^2.8.6"
djangorestframework-simplejwt = "^4.5.0"
whitenoise = "^5.2.0"
django-phonenumber-field = "^5.0.0"
phonenumbers = "^8.12.12"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
